0.4.7.4 - fixed visual bug caused by changing the device orientation - clock added - sundered model into abstract xor concrete, i.e. no abstract classes

+++ Assets/Model/ClockInfo.cs
@@ -2,21 +2,21 @@ namespace Assets.Model
 {
     public sealed class ClockInfo
     {
-        private readonly int _hour;
-        private readonly int _minute;
-        private readonly int _second;
+        private readonly float _hour;
+        private readonly float _minute;
+        private readonly float _second;
 
-        public ClockInfo(int hour, int minute, int second)
+        public ClockInfo(float hour, float minute, float second)
         {
             _hour = hour;
             _minute = minute;
             _second = second;
         }
 
-        public int Hour => _hour;
+        public float Hour => _hour;
 
-        public int Minute => _minute;
+        public float Minute => _minute;
 
-        public int Second => _second;
+        public float Second => _second;
     }
 }
